requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-environ==0.9.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.0
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1